Use fp16

+++ stable_diffusion/ldm/modules/encoders/modules.py
@@ -138,8 +138,8 @@ class FrozenCLIPEmbedder(AbstractEncoder):
     """Uses the CLIP transformer encoder for text (from Hugging Face)"""
     def __init__(self, version="openai/clip-vit-large-patch14", device="cuda", max_length=77):
         super().__init__()
-        self.tokenizer = CLIPTokenizer.from_pretrained(version)
-        self.transformer = CLIPTextModel.from_pretrained(version)
+        self.tokenizer = CLIPTokenizer.from_pretrained(version, torch_dtype=torch.float16)
+        self.transformer = CLIPTextModel.from_pretrained(version, torch_dtype=torch.float16)
         self.device = device
         self.max_length = max_length
         self.freeze()
